Now, kommit check if the hook is installed?
It is possible to forget to apply hook. kommit keeps continue to work
but when you are ready to make the real commit, messages are not pop!

Right now, it doesn't matter if you installed the hook or not.

Your first message addition is handling this. All you need is to add
a first message via: `kommit -m` :)